LayoutTests/http/tests/websocket/tests/hybi/handshake-fail-by-invalid-http-version-expected.txt

Test whether WebSocket handshake fails if the server sends status line with invalid http version.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
